Comparing signed and unsigned might be a problem.

+++ net/sched/sch_deltic-boroshne.c
@@ -598,9 +598,9 @@ static struct sk_buff* boroshne_dequeue(struct Qdisc *sch)
 		WARN_ON(q->bulk_flows && !q->bulk_bklg);
 		WARN_ON(q->hogg_flows && !q->hogg_bklg);
 
-		while(	(quik_blocked && quik_inc < -q->quik_deficit) ||
-				(bulk_blocked && bulk_inc < -q->bulk_deficit) ||
-				(hogg_blocked && hogg_inc < -q->hogg_deficit) )
+		while(	(quik_blocked && ((s32) quik_inc) < -q->quik_deficit) ||
+				(bulk_blocked && ((s32) bulk_inc) < -q->bulk_deficit) ||
+				(hogg_blocked && ((s32) hogg_inc) < -q->hogg_deficit) )
 		{
 			quik_inc *= 2;
 			bulk_inc *= 2;
